ControlNet-v1-1-nightly/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.0
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
attrs==23.2.0
av==11.0.0
basicsr==1.4.2
blinker==1.7.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
docker-pycreds==0.4.0
einops==0.7.0
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==7.0.1
importlib-resources==6.1.1
invisible-watermark==0.2.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.3
lightning-utilities==0.10.1
lmdb==1.4.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.16.3
open-clip-torch==2.24.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.9.12
packaging==23.2
pandas==2.2.0
pillow==10.2.0
portalocker==2.8.2
prettytable==3.6.0
proglog==0.1.10
protobuf==4.23.4
psutil==5.9.8
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.20.0
scikit-learn==1.4.0
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.35.1
streamlit==1.30.0
streamlit-drawable-canvas==0.8.0
sympy==1.12
tabulate==0.9.0
tb-nightly==2.16.0a20240122
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
test_tube==0.7.5
tf_keras-nightly==2.16.0.dev2024012210
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.6.12
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchdata==0.7.1
torchmetrics==0.11.4
torchtext==0.16.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
transformers==4.37.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.26.0
validators==0.22.0
wandb==0.16.2
wcwidth==0.2.13
webdataset==0.2.5
websockets==11.0.3
Werkzeug==3.0.1
yacs==0.1.8
yapf==0.32.0
yarl==1.9.4
zipp==3.17.0
